requirements.txt
pyairtable==1.1.0
PyPDF2==3.0.1
google-generativeai
numpy
pandas